+++ composables/ving/useRestVersion.mjs
@@ -1,4 +1,4 @@
-export const restVersion = () => {
+export const useRestVersion = () => {
     const config = useRuntimeConfig();
     return config.public.rest.version;
 }