+++ pages/user/admin/[id].vue
@@ -103,8 +103,8 @@ const notify = useNotifyStore();
 const id = route.params.id.toString();
 const user = useVingRecord({
     id,
-    fetchApi: `/api/${restVersion()}/user/${id}`,
-    createApi: `/api/${restVersion()}/user`,
+    fetchApi: `/api/${useRestVersion()}/user/${id}`,
+    createApi: `/api/${useRestVersion()}/user`,
     query: { includeMeta: true, includeOptions: true },
     onUpdate() {
         notify.success('Updated user.');